refactor(package directory)
- GameResult가 삭제됨에 따라 불필요하게 세분화 된 gameresult 패키지 폐기 후 game 패키지로 통합
