+++ README.md
@@ -241,4 +241,4 @@
 - [X] HorizontalLocation에서 움직이는 메서드 인자 개선
 - [X] 사용하지 않는 메서드 제거
 - [X] 쓸모없는 getter 제거하고 객체 비교하기
-- [ ] GameResults를 Map으로 개선
+- [X] GameResults를 Map으로 개선
